moved dialog to separated file
